feat: add formatting support and refactor auto-correction logic
